feat(tree-grid): fixing grid modules #2530
